Move updatecheck to pkg
