+++ cmd/http.go
@@ -11,6 +11,7 @@ import (
 	lm "github.com/loophole/cli/internal/app/loophole/models"
 	"github.com/loophole/cli/internal/pkg/communication"
 	"github.com/loophole/cli/internal/pkg/token"
+	"github.com/loophole/cli/internal/pkg/updatecheck"
 
 	"github.com/spf13/cobra"
 )
@@ -29,7 +30,7 @@ To expose port running on some local host e.g. 192.168.1.20 use 'loophole http <
 		idToken := token.GetIdToken()
 		communication.ApplicationStart(loggedIn, idToken)
 
-		loophole.CheckVersion()
+		updatecheck.CheckVersion()
 
 		localEndpointSpecs.Host = "127.0.0.1"
 		if len(args) > 1 {
