+++ cmd/path.go
@@ -9,6 +9,7 @@ import (
 	lm "github.com/loophole/cli/internal/app/loophole/models"
 	"github.com/loophole/cli/internal/pkg/communication"
 	"github.com/loophole/cli/internal/pkg/token"
+	"github.com/loophole/cli/internal/pkg/updatecheck"
 	"github.com/spf13/cobra"
 )
 
@@ -26,7 +27,7 @@ To expose local directory (e.g. /data/my-data) simply use 'loophole path /data/m
 		idToken := token.GetIdToken()
 		communication.ApplicationStart(loggedIn, idToken)
 
-		loophole.CheckVersion()
+		updatecheck.CheckVersion()
 
 		dirEndpointSpecs.Path = args[0]
 		quitChannel := make(chan bool)
