+++ internal/app/loophole/loophole.go
@@ -5,10 +5,8 @@ import (
 	"io"
 	"net"
 	"net/http"
-	"runtime"
 	"time"
 
-	"github.com/blang/semver/v4"
 	"github.com/loophole/cli/config"
 	lm "github.com/loophole/cli/internal/app/loophole/models"
 	"github.com/loophole/cli/internal/pkg/apiclient"
@@ -16,10 +14,6 @@ import (
 	"github.com/loophole/cli/internal/pkg/httpserver"
 	"github.com/loophole/cli/internal/pkg/keys"
 	"github.com/loophole/cli/internal/pkg/urlmaker"
-	"github.com/mitchellh/go-homedir"
-	"github.com/ncruces/zenity"
-	"github.com/pkg/browser"
-	"github.com/spf13/viper"
 	"golang.org/x/crypto/ssh"
 )
 
@@ -39,92 +33,6 @@ const (
 var remoteEndpoint = lm.Endpoint{
 	Host: "127.0.0.1",
 	Port: 80,
-}
-
-func CheckVersion() {
-	availableVersion, err := apiclient.GetLatestAvailableVersion()
-	if err != nil {
-		communication.Debug("There was a problem obtaining info response, skipping further checking")
-		return
-	}
-	currentVersionParsed, err := semver.Make(config.Config.Version)
-	if err != nil {
-		communication.Debug(fmt.Sprintf("Cannot parse current version '%s' as semver version, skipping further checking", config.Config.Version))
-		return
-	}
-	availableVersionParsed, err := semver.Make(availableVersion.Version)
-	if err != nil {
-		communication.Debug(fmt.Sprintf("Cannot parse available version '%s' as semver version, skipping further checking", availableVersion))
-		return
-	}
-	if currentVersionParsed.LT(availableVersionParsed) {
-		if config.Config.ClientMode == "cli" {
-			communication.NewVersionAvailable(availableVersion.Version)
-		} else {
-			remind, _ := remindAgainCheck()
-			if !remind {
-				return
-			}
-			dlLink := getDownloadLink(availableVersion.Version)
-			response, _ := zenity.Question(fmt.Sprintf("A new version is available for you at \n%s \nDo you want to open the link in your browser now?", dlLink), zenity.NoWrap(), zenity.Title("New version available!"))
-			if response {
-				browser.OpenURL(dlLink)
-			}
-		}
-	}
-}
-
-func getDownloadLink(availableVersion string) string {
-	archiveExt := ".tar.gz"
-	arch := runtime.GOARCH
-	if arch == "windows" {
-		archiveExt = ".zip"
-	} else if arch == "darwin" {
-		arch = "macos"
-	}
-	if arch == "amd64" {
-		arch = "64bit"
-	} else if arch == "386" {
-		arch = "32bit"
-	} else {
-		communication.Error("There was an error detecting your system architecture.") //if arch is unexpected, only link to the release page
-		return fmt.Sprintf("https://github.com/loophole/cli/releases/tag/%s", availableVersion)
-	}
-	res := fmt.Sprintf("https://github.com/loophole/cli/releases/download/%s/loophole-desktop_%s_%s_%s%s", availableVersion, availableVersion, runtime.GOOS, arch, archiveExt)
-	fmt.Println(res)
-	return res
-}
-
-func remindAgainCheck() (bool, error) { //TODO: Error handling, moving this function to a more appropriate file (probably to config pkg)
-	home, err := homedir.Dir()
-	if err != nil {
-		return true, err
-	}
-
-	layout := "2006-02-01"                                        //golangs arcane time format string
-	viper.SetDefault("last-reminder", time.Time{}.Format(layout)) //zero value for time
-	viper.SetConfigName("config")                                 // name of config file (without extension)
-	viper.SetConfigType("json")                                   // REQUIRED if the config file does not have the extension in the name
-	viper.AddConfigPath(fmt.Sprintf("%s/.loophole/", home))
-	if err := viper.ReadInConfig(); err != nil {
-		if _, ok := err.(viper.ConfigFileNotFoundError); ok { //create a config if none exist yet
-			viper.WriteConfigAs(fmt.Sprintf("%s/.loophole/config.json", home))
-		} else {
-			return true, err
-		}
-	}
-
-	t, err := time.Parse(layout, fmt.Sprintf("%v", viper.Get("last-reminder")))
-	if err != nil {
-		return true, err
-	}
-	if (t.Year() < time.Now().Year()) || (t.YearDay() < time.Now().YearDay()) { //check if reminder has been done today
-		viper.Set("last-reminder", time.Now().Format(layout))
-		viper.WriteConfigAs("/home/work/.loophole/config.json")
-		return true, nil
-	}
-
-	return false, nil
 }
 
 func handleClient(tunnelID string, client net.Conn, local net.Conn) {
